Move integration list out of pages directory
This fixes a build issue with docusaurus. It was erroring because
it couldn't generate a page for "/integrations/all" when `all.ts`
was in the pages directory.
